progress

+++ src/lib/components/file-uploader/_partials/preloadedFileRow.svelte
@@ -1,5 +1,5 @@
 <script>
-    import Button from "$lib/components/ui/button/index";
+    import { Button } from "$lib/components/ui/button/index.js";
     import { Pencil, X } from "lucide-svelte";
     import MimeType from "../mimeType.svelte";
     import { fade, slide } from "svelte/transition";
@@ -7,15 +7,14 @@
 
     export let preloadedFiles;
     export let index;
+    export let deleteEndpoint = "api/s3/upload-file";
 
     const dispatch = createEventDispatcher();
 
-    const removeFile = async (guid = "", toRemoveIndex) => {
-        const deleteResult = await fetch("api/s3/upload-file", { method: "DELETE", body: JSON.stringify({ guid }) });
-        dispatch("deleted", { toRemoveIndex: index });
+    const removeFile = async (guid = "") => {
+        const deleteResult = await fetch(deleteEndpoint, { method: "DELETE", body: JSON.stringify({ guid }) });
+        if (deleteResult.ok) dispatch("deleted", { toRemoveIndex: index });
     };
-
-    console.log(preloadedFiles);
 </script>
 
 <div class="mt-1 flex w-full justify-between overflow-hidden rounded bg-gray-200" in:slide out:fade>
@@ -26,10 +25,10 @@
         <a href={preloadedFiles[index].url} target="_blank" class="truncate">{preloadedFiles[index].displayName} </a>
     </span>
     <span class="flex flex-nowrap items-center gap-1 bg-red-200 px-2">
-        <Button variant="outline" size="inline-icon" on:click={() => removeFile(preloadedFiles[index].guid, index)}>
+        <Button variant="outline" size="inline-icon" on:click={() => removeFile(preloadedFiles[index].guid)}>
             <Pencil class="h-4 w-4" />
         </Button>
-        <Button variant="destructive-outline" size="inline-icon" on:click={() => removeFile(preloadedFiles[index].guid, index)}>
+        <Button variant="destructive-outline" size="inline-icon" on:click={() => removeFile(preloadedFiles[index].guid)}>
             <X class="h-4 w-4" />
         </Button>
     </span>
